backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
openai==1.51.0
boto3==1.35.0
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
scikit-learn==1.5.2
snowflake-connector-python==3.12.0
pgvector==0.3.2
psycopg2-binary==2.9.9
